+++ packages/azure/src/azure-openai-provider.test.ts
@@ -1332,7 +1332,7 @@ describe('responses', () => {
           {
             "id": "msg_67c9a81dea8c8190b79651a2b3adf91e",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "itemId": "msg_67c9a81dea8c8190b79651a2b3adf91e",
               },
             },
@@ -1351,7 +1351,7 @@ describe('responses', () => {
           {
             "id": "msg_67c9a8787f4c8190b49c858d4c1cf20c",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "itemId": "msg_67c9a8787f4c8190b49c858d4c1cf20c",
               },
             },
@@ -1360,7 +1360,7 @@ describe('responses', () => {
           {
             "finishReason": "stop",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "responseId": "resp_67c9a81b6a048190a9ee441c5755a4e8",
               },
             },
@@ -1436,7 +1436,7 @@ describe('responses', () => {
           {
             "input": "{}",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "itemId": "fc_67cb13a838088190be08eb3927c87501",
               },
             },
@@ -1481,7 +1481,7 @@ describe('responses', () => {
           {
             "input": "{"location":"Rome"}",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "itemId": "fc_67cb13a858f081908a600343fa040f47",
               },
             },
@@ -1492,7 +1492,7 @@ describe('responses', () => {
           {
             "finishReason": "tool-calls",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "responseId": "resp_67cb13a755c08190acbe3839a49632fc",
               },
             },
@@ -1541,7 +1541,7 @@ describe('responses', () => {
             "id": "id-0",
             "mediaType": "text/plain",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "fileId": "assistant-YRcoCqn3Fo2K4JgraG",
                 "filename": "resource1.json",
                 "index": 145,
@@ -1557,7 +1557,7 @@ describe('responses', () => {
             "id": "id-1",
             "mediaType": "text/plain",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "fileId": "assistant-YRcoCqn3Fo2K4JgraG",
                 "filename": "resource1.json",
                 "index": 192,
@@ -1571,7 +1571,7 @@ describe('responses', () => {
           {
             "id": "msg_456",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "annotations": [
                   {
                     "file_id": "assistant-YRcoCqn3Fo2K4JgraG",
@@ -1594,7 +1594,7 @@ describe('responses', () => {
           {
             "finishReason": "stop",
             "providerMetadata": {
-              "openai": {
+              "azure": {
                 "responseId": null,
               },
             },