+++ packages/openai/src/responses/openai-responses-language-model.ts
@@ -364,6 +364,8 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
       modelId: this.modelId,
     });
 
+    const providerKey = this.config.provider.replace('.responses', ''); // can be 'openai' or 'azure'. provider is 'openai.responses' or 'azure.responses'.
+
     const {
       responseHeaders,
       value: response,
@@ -412,7 +414,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
               type: 'reasoning' as const,
               text: summary.text,
               providerMetadata: {
-                openai: {
+                [providerKey]: {
                   itemId: part.id,
                   reasoningEncryptedContent: part.encrypted_content ?? null,
                 },
@@ -452,7 +454,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
               action: part.action,
             } satisfies InferSchema<typeof localShellInputSchema>),
             providerMetadata: {
-              openai: {
+              [providerKey]: {
                 itemId: part.id,
               },
             },
@@ -481,7 +483,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
               type: 'text',
               text: contentPart.text,
               providerMetadata: {
-                openai: providerMetadata,
+                [providerKey]: providerMetadata,
               },
             });
 
@@ -503,7 +505,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: annotation.filename,
                   filename: annotation.filename,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: annotation.type,
                       fileId: annotation.file_id,
                       filename: annotation.filename,
@@ -520,7 +522,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: annotation.filename,
                   filename: annotation.filename,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: annotation.type,
                       fileId: annotation.file_id,
                       containerId: annotation.container_id,
@@ -537,7 +539,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: annotation.file_id,
                   filename: annotation.file_id,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: annotation.type,
                       fileId: annotation.file_id,
                       index: annotation.index,
@@ -560,7 +562,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
             toolName: part.name,
             input: part.arguments,
             providerMetadata: {
-              openai: {
+              [providerKey]: {
                 itemId: part.id,
               },
             },
@@ -745,15 +747,15 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
     }
 
     const providerMetadata: SharedV3ProviderMetadata = {
-      openai: { responseId: response.id },
+      [providerKey]: { responseId: response.id },
     };
 
     if (logprobs.length > 0) {
-      providerMetadata.openai.logprobs = logprobs;
+      providerMetadata[providerKey].logprobs = logprobs;
     }
 
     if (typeof response.service_tier === 'string') {
-      providerMetadata.openai.serviceTier = response.service_tier;
+      providerMetadata[providerKey].serviceTier = response.service_tier;
     }
 
     const usage = response.usage!; // defined when there is no error
@@ -815,6 +817,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
     });
 
     const self = this;
+    const providerKey = this.config.provider.replace('.responses', ''); // can be 'openai' or 'azure'. provider is 'openai.responses' or 'azure.responses'.
 
     let finishReason: LanguageModelV3FinishReason = 'unknown';
     const usage: LanguageModelV3Usage = {
@@ -986,7 +989,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   type: 'text-start',
                   id: value.item.id,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       itemId: value.item.id,
                     },
                   },
@@ -1004,7 +1007,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   type: 'reasoning-start',
                   id: `${value.item.id}:0`,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       itemId: value.item.id,
                       reasoningEncryptedContent:
                         value.item.encrypted_content ?? null,
@@ -1031,7 +1034,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   toolName: value.item.name,
                   input: value.item.arguments,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       itemId: value.item.id,
                     },
                   },
@@ -1186,7 +1189,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                     },
                   } satisfies InferSchema<typeof localShellInputSchema>),
                   providerMetadata: {
-                    openai: { itemId: value.item.id },
+                    [providerKey]: { itemId: value.item.id },
                   },
                 });
               } else if (value.item.type === 'reasoning') {
@@ -1208,7 +1211,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                     type: 'reasoning-end',
                     id: `${value.item.id}:${summaryIndex}`,
                     providerMetadata: {
-                      openai: {
+                      [providerKey]: {
                         itemId: value.item.id,
                         reasoningEncryptedContent:
                           value.item.encrypted_content ?? null,
@@ -1315,7 +1318,9 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                     controller.enqueue({
                       type: 'reasoning-end',
                       id: `${value.item_id}:${summaryIndex}`,
-                      providerMetadata: { openai: { itemId: value.item_id } },
+                      providerMetadata: {
+                        [providerKey]: { itemId: value.item_id },
+                      },
                     });
                     activeReasoningPart.summaryParts[summaryIndex] =
                       'concluded';
@@ -1326,7 +1331,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   type: 'reasoning-start',
                   id: `${value.item_id}:${value.summary_index}`,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       itemId: value.item_id,
                       reasoningEncryptedContent:
                         activeReasoning[value.item_id]?.encryptedContent ??
@@ -1341,7 +1346,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                 id: `${value.item_id}:${value.summary_index}`,
                 delta: value.delta,
                 providerMetadata: {
-                  openai: {
+                  [providerKey]: {
                     itemId: value.item_id,
                   },
                 },
@@ -1354,7 +1359,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   type: 'reasoning-end',
                   id: `${value.item_id}:${value.summary_index}`,
                   providerMetadata: {
-                    openai: { itemId: value.item_id },
+                    [providerKey]: { itemId: value.item_id },
                   },
                 });
 
@@ -1407,7 +1412,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: value.annotation.filename,
                   filename: value.annotation.filename,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: value.annotation.type,
                       fileId: value.annotation.file_id,
                       filename: value.annotation.filename,
@@ -1424,7 +1429,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: value.annotation.filename,
                   filename: value.annotation.filename,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: value.annotation.type,
                       fileId: value.annotation.file_id,
                       containerId: value.annotation.container_id,
@@ -1441,7 +1446,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                   title: value.annotation.file_id,
                   filename: value.annotation.file_id,
                   providerMetadata: {
-                    openai: {
+                    [providerKey]: {
                       type: value.annotation.type,
                       fileId: value.annotation.file_id,
                       index: value.annotation.index,
@@ -1457,7 +1462,7 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
                 type: 'text-end',
                 id: value.item.id,
                 providerMetadata: {
-                  openai: {
+                  [providerKey]: {
                     itemId: value.item.id,
                     ...(ongoingAnnotations.length > 0 && {
                       annotations: ongoingAnnotations,
@@ -1472,17 +1477,17 @@ export class OpenAIResponsesLanguageModel implements LanguageModelV3 {
 
           flush(controller) {
             const providerMetadata: SharedV3ProviderMetadata = {
-              openai: {
+              [providerKey]: {
                 responseId,
               },
             };
 
             if (logprobs.length > 0) {
-              providerMetadata.openai.logprobs = logprobs;
+              providerMetadata[providerKey].logprobs = logprobs;
             }
 
             if (serviceTier !== undefined) {
-              providerMetadata.openai.serviceTier = serviceTier;
+              providerMetadata[providerKey].serviceTier = serviceTier;
             }
 
             controller.enqueue({