+++ packages/lmnt/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/lmnt
 
+## 2.0.0-beta.34
+
+### Patch Changes
+
+- Updated dependencies [db913bd]
+  - @ai-sdk/provider@3.0.0-beta.17
+  - @ai-sdk/provider-utils@4.0.0-beta.34
+
 ## 2.0.0-beta.33
 
 ### Patch Changes